+++ src/SSCM.Core/MappingDataRepositoryDefault.cs
@@ -1,8 +1,7 @@
-using SSCM.Core;
+namespace SSCM.Core;
 
-namespace SSCM.StarCitizen;
-
-public class MappingDataRepository : IMappingDataRepository<MappingData>
+public class MappingDataRepositoryDefault<TData> : IMappingDataRepository<TData>
+    where TData : class, new()
 {
     public event Action<string> StandardOutput = delegate {};
     public event Action<string> WarningOutput = delegate {};
@@ -15,7 +14,7 @@ public class MappingDataRepository : IMappingDataRepository<MappingData>
     private readonly IPlatform _platform;
     private readonly string _mappingDataSaveDir;
 
-    public MappingDataRepository(IPlatform platform, string mappingDataSavePath, string backupFilenameFormat)
+    public MappingDataRepositoryDefault(IPlatform platform, string mappingDataSavePath, string backupFilenameFormat)
     {
         this._platform = platform;
         this.MappingDataSavePath = mappingDataSavePath;
@@ -23,20 +22,20 @@ public class MappingDataRepository : IMappingDataRepository<MappingData>
         this.BackupFilenameFormat = backupFilenameFormat;
     }
 
-    public MappingData CreateNew() => new MappingData();
+    public TData CreateNew() => new TData();
 
-    public async Task<MappingData?> Load(string? saveFilePath = null)
+    public async Task<TData?> Load(string? saveFilePath = null)
     {
         if (string.IsNullOrWhiteSpace(saveFilePath))
         {
             saveFilePath = this.MappingDataSavePath;
         }
 
-        var serializer = new DataSerializer<MappingData>(saveFilePath);
+        var serializer = new DataSerializer<TData>(saveFilePath);
         return await serializer.Read();
     }
 
-    public async Task Save(MappingData data, string? saveFilePath = null)
+    public async Task Save(TData data, string? saveFilePath = null)
     {
         if (data == null)
         {
@@ -50,7 +49,7 @@ public class MappingDataRepository : IMappingDataRepository<MappingData>
 
         var saveDir = new FileInfo(saveFilePath).DirectoryName;
         System.IO.Directory.CreateDirectory(saveDir);
-        var serializer = new DataSerializer<MappingData>(saveFilePath);
+        var serializer = new DataSerializer<TData>(saveFilePath);
         await serializer.Write(data);
         this.StandardOutput($"Mappings saved to [{this.MappingDataSavePath}].");
     }