+++ src/SSCM.Elite/MappingData.cs
@@ -1,0 +1,5 @@
+namespace SSCM.Elite;
+
+public class MappingData
+{
+}