+++ src/SSCM.Elite/MappingMergeResult.cs
@@ -1,0 +1,17 @@
+using SSCM.Core;
+using System.Text;
+
+namespace SSCM.Elite;
+
+public class MappingMergeResult : MappingMergeResultBase<MappingData>
+{
+    public MappingMergeResult(MappingData current, MappingData updated) : base(current, updated)
+    {
+    }
+
+    public override string ToString()
+    {
+        var sb = new StringBuilder();
+        return sb.ToString();
+    }
+}